third_party/blink/web_tests/external/wpt/css/css-scroll-anchoring/dirty-contents-reselect-anchor.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Scroll anchor is re-selected after adjustment if there are dirty descendants at selection time
  assert_equals: Scroll position should've been preserved expected 1207 but got 0
Harness: the test ran to completion.
